requirements.txt
shapely
pyproj
Pillow
typer
numpy
trimesh
